complete_requirements.txt
altair==4.2.0
asgiref==3.5.2
attrs==22.1.0
bert-extractive-summarizer==0.10.1
blinker==1.5
blis==0.7.9
cachetools==5.2.0
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
confection==0.0.3
cymem==2.0.7
decorator==5.1.1
Django==4.1.4
djangorestframework==3.14.0
entrypoints==0.4
ffmpeg-python==0.2.0
filelock==3.8.2
future==0.18.2
gitdb==4.0.10
GitPython==3.1.29
huggingface-hub==0.11.1
idna==3.4
importlib-metadata==5.1.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
langcodes==3.3.0
MarkupSafe==2.1.1
more-itertools==9.0.0
msgpack==0.6.2
murmurhash==1.0.9
numpy==1.23.5
packaging==22.0
pandas==1.5.2
pathy==0.10.1
Pillow==9.3.0
plac==0.9.6
preshed==3.0.8
protobuf==3.20.3
pyarrow==10.0.1
pydantic==1.10.2
pydeck==0.8.0
Pygments==2.13.0
Pympler==1.0.1
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2017.4.5
requests==2.28.1
rich==12.6.0
scikit-learn==1.2.0
scipy==1.9.3
semver==2.13.0
six==1.16.0
smart-open==6.2.0
smmap==5.0.0
spacy==3.4.3
spacy-legacy==3.0.10
spacy-loggers==1.0.4
sqlparse==0.4.3
srsly==2.4.5
streamlit==1.15.2
thinc==8.1.5
threadpoolctl==3.1.0
tokenizers==0.13.2
toml==0.10.2
toolz==0.12.0
torch==1.13.0
tornado==6.2
tqdm==4.64.1
transformers==4.25.1
typer==0.7.0
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
validators==0.20.0
wasabi==0.10.1
whisper @ git+https://github.com/openai/whisper.git@0b5dcfdef7ec04250b76e13f1630e32b0935ce76
wrapt==1.10.11
youtube-dl==2021.12.17
zipp==3.11.0
